Move data models to `app-data/data.models`
