+++ app/shared/app-data/src/commonMain/kotlin/data/models/EpisodeHistory.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
+package me.him188.ani.app.data.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class EpisodeHistory(
+    val episodeId: Int,
+    val positionMillis: Long
+)
