+++ app/shared/app-data/src/commonMain/kotlin/data/repository/EpisodePlayHistoryRepository.kt
@@ -1,3 +1,12 @@
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
 package me.him188.ani.app.data.repository
 
 import androidx.datastore.core.DataStore
@@ -5,6 +14,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
 import kotlinx.serialization.Serializable
+import me.him188.ani.app.data.models.EpisodeHistory
 import me.him188.ani.utils.logging.info
 import me.him188.ani.utils.logging.logger
 
@@ -16,12 +26,6 @@ interface EpisodePlayHistoryRepository : Repository {
     suspend fun saveOrUpdate(episodeId: Int, positionMillis: Long)
     suspend fun getPositionMillisByEpisodeId(episodeId: Int): Long?
 }
-
-@Serializable
-data class EpisodeHistory(
-    val episodeId: Int,
-    val positionMillis: Long
-)
 
 @Serializable
 data class EpisodeHistories(
