+++ app/shared/src/commonMain/kotlin/ui/subject/collection/CollectionPage.kt
@@ -67,10 +67,10 @@ import androidx.paging.PagingData
 import androidx.paging.compose.collectAsLazyPagingItemsWithLifecycle
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
+import me.him188.ani.app.data.models.SubjectCollectionInfo
 import me.him188.ani.app.data.models.UserInfo
 import me.him188.ani.app.data.models.subject.SubjectCollectionCounts
 import me.him188.ani.app.data.repository.CollectionsFilterQuery
-import me.him188.ani.app.data.repository.SubjectCollectionInfo
 import me.him188.ani.app.domain.session.AuthState
 import me.him188.ani.app.navigation.LocalNavigator
 import me.him188.ani.app.ui.adaptive.AniTopAppBar
