+++ app/shared/src/commonMain/kotlin/ui/subject/details/SubjectDetailsState.kt
@@ -1,3 +1,12 @@
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
 package me.him188.ani.app.ui.subject.details
 
 import androidx.compose.runtime.Stable
@@ -26,7 +35,7 @@ class SubjectDetailsState(
     relatedSubjectsState: State<List<RelatedSubjectInfo>?>,
     val airingLabelState: AiringLabelState,
 ) {
-    val info by derivedStateOf { subjectInfoState.value ?: SubjectInfo.Empty }
+    val info by derivedStateOf { subjectInfoState.value ?: me.him188.ani.app.data.models.SubjectInfo.Empty }
 
     val coverImageUrl by derivedStateOf { info.imageLarge }
 
@@ -49,7 +58,7 @@ class SubjectDetailsState(
      */
     val isLoading: Boolean by derivedStateOf {
         selfCollectionTypeOrNull == null
-                || charactersOrNull == null || info === SubjectInfo.Empty
+                || charactersOrNull == null || info === me.him188.ani.app.data.models.SubjectInfo.Empty
                 || personsOrNull == null
                 || relatedSubjectsOrNull == null
                 || airingLabelState.isLoading
