+++ app/shared/src/desktopTest/kotlin/ui/subject/cache/EpisodeCacheStateTest.kt
@@ -20,7 +20,6 @@ import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runCurrent
 import kotlinx.coroutines.test.runTest
 import me.him188.ani.app.data.models.episode.EpisodeInfo
-import me.him188.ani.app.data.models.subject.SubjectInfo
 import me.him188.ani.app.domain.media.TestMediaList
 import me.him188.ani.app.domain.media.cache.EpisodeCacheStatus
 import me.him188.ani.app.domain.media.cache.requester.CacheRequestStage
@@ -31,9 +30,9 @@ import me.him188.ani.app.domain.media.cache.storage.TestMediaCacheStorage
 import me.him188.ani.app.domain.media.fetch.MediaFetcherConfig
 import me.him188.ani.app.domain.media.fetch.MediaSourceMediaFetcher
 import me.him188.ani.app.domain.media.framework.TestMediaSelector
-import me.him188.ani.app.domain.mediasource.instance.createTestMediaSourceInstance
 import me.him188.ani.app.domain.media.selector.MediaSelector
 import me.him188.ani.app.domain.media.selector.MediaSelectorFactory
+import me.him188.ani.app.domain.mediasource.instance.createTestMediaSourceInstance
 import me.him188.ani.app.ui.foundation.produceState
 import me.him188.ani.app.ui.framework.runComposeStateTest
 import me.him188.ani.app.ui.framework.takeSnapshot
@@ -206,7 +205,7 @@ class EpisodeCacheStateTest {
     fun `showProgressIndicator is true when SelectMedia`() = runComposeStateTest {
         val state = createEpisodeCacheState()
         state.cacheRequester
-            .request(EpisodeCacheRequest(SubjectInfo.Empty, EpisodeInfo(1)))
+            .request(EpisodeCacheRequest(me.him188.ani.app.data.models.SubjectInfo.Empty, EpisodeInfo(1)))
 
         takeSnapshot()
         assertEquals(true, state.showProgressIndicator)
@@ -216,7 +215,7 @@ class EpisodeCacheStateTest {
     fun `showProgressIndicator is true when SelectStorage`() = runComposeStateTest {
         val state = createEpisodeCacheState()
         state.cacheRequester
-            .request(EpisodeCacheRequest(SubjectInfo.Empty, EpisodeInfo(1)))
+            .request(EpisodeCacheRequest(me.him188.ani.app.data.models.SubjectInfo.Empty, EpisodeInfo(1)))
             .select(TestMediaList[0])
 
         takeSnapshot()
@@ -227,7 +226,7 @@ class EpisodeCacheStateTest {
     fun `showProgressIndicator is false when Done`() = runComposeStateTest {
         val state = createEpisodeCacheState()
         state.cacheRequester
-            .request(EpisodeCacheRequest(SubjectInfo.Empty, EpisodeInfo(1)))
+            .request(EpisodeCacheRequest(me.him188.ani.app.data.models.SubjectInfo.Empty, EpisodeInfo(1)))
             .select(TestMediaList[0])
             .trySelectSingle()
 
@@ -250,7 +249,7 @@ class EpisodeCacheStateTest {
     fun `currentSelectMediaTask is not null when SelectMedia`() = runComposeStateTest {
         val state = createEpisodeCacheState()
         state.cacheRequester
-            .request(EpisodeCacheRequest(SubjectInfo.Empty, EpisodeInfo(1)))
+            .request(EpisodeCacheRequest(me.him188.ani.app.data.models.SubjectInfo.Empty, EpisodeInfo(1)))
 
         takeSnapshot()
 
@@ -261,7 +260,7 @@ class EpisodeCacheStateTest {
     fun `currentStage is Done when Done`() = runComposeStateTest {
         val state = createEpisodeCacheState()
         state.cacheRequester
-            .request(EpisodeCacheRequest(SubjectInfo.Empty, EpisodeInfo(1)))
+            .request(EpisodeCacheRequest(me.him188.ani.app.data.models.SubjectInfo.Empty, EpisodeInfo(1)))
             .select(TestMediaList[0])
             .trySelectSingle()
